code/book/fastapi-t/fastcode/blog_app/requirements.txt
aiomysql==0.1.1
fastapi==0.78.0
passlib==1.7.4
pydantic==1.9.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.19.1
uvicorn==0.17.6
